requirements.txt
git+https://github.com/openai/swarm.git
ollama
lancedb
PyYAML
torch
numpy
python-dotenv